# mobile_shop_api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Integration Swagger in Laravel with Sanctum Auth Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get All Category",
                "operationId": "getCategories",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sort by column: 0-id, 1-name, 2-sort_no, 3-home, 4-image, 5-created_at",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": "0",
                            "enum": [
                                0,
                                1,
                                2,
                                3,
                                4,
                                5
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "sort by order: 0-ASC, 1-DESC",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": "0",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "sort by paginate page: 0-10, 1-25, 2-50, 3-100",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": "0",
                            "enum": [
                                0,
                                1,
                                2,
                                3
                            ]
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "start date to filter (dd/mm/yyyy)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "end date to filter (dd/mm/yyyy)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Add Category With Form Data",
                "operationId": "storeCategory",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "home",
                                    "sort_no",
                                    "image"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "sort_no": {
                                        "type": "integer",
                                        "default": "0"
                                    },
                                    "home": {
                                        "description": "Show in homepage => 0: False, 1: True",
                                        "type": "integer",
                                        "default": "0",
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Find Category By ID",
                "operationId": "showCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Update Category With Form Data",
                "operationId": "updateCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "home",
                                    "sort_no",
                                    "_method"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "sort_no": {
                                        "type": "integer",
                                        "default": "0"
                                    },
                                    "home": {
                                        "description": "Show in homepage => 0: False, 1: True",
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ]
                                    },
                                    "image": {
                                        "type": "file"
                                    },
                                    "_method": {
                                        "type": "string",
                                        "default": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Method not allow"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete Category By ID",
                "operationId": "destroyCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories/excel/export": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Export Excel Category",
                "operationId": "exportCategory",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sort by column: 0-id, 1-name, 2-sort_no, 3-home, 4-image, 5-created_at",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": "0",
                            "enum": [
                                0,
                                1,
                                2,
                                3,
                                4,
                                5
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "sort by order: 0-ASC, 1-DESC",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": "0",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "sort by paginate page: 0-10, 1-25, 2-50, 3-100",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": "0",
                            "enum": [
                                0,
                                1,
                                2,
                                3
                            ]
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "start date to filter (dd/mm/yyyy)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "end date to filter (dd/mm/yyyy)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "title": "Category",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Điện thoại"
                    },
                    "sort_no": {
                        "type": "integer",
                        "example": "1"
                    },
                    "home": {
                        "description": "Show in homepage => 0: False, 1: True",
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ],
                        "example": "0"
                    },
                    "image": {
                        "type": "string",
                        "example": "dien-thoai.png"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Category"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}